rescan-scsi-bus.sh: fix for github.com/doug-gilbert/sg3_utils/issues/46
The uses of

   sg_turs "$sg_turs_opt" /dev/$SGDEV >/dev/null 2>&1

are incorrect if there are no options, as it passes the empty string as the
first argument to sg_turs which treats it as the device positional
argument. The quotes should be removed (though I don't know if that affects
the case where there is an argument).

git-svn-id: https://svn.bingwo.ca/repos/sg3_utils/trunk@1076 6180dd3e-e324-4e3e-922d-17de1ae2f315

In the worst case, this may lead to file systems going read only because
a SCSI device is erroneously considered offline.

Fixes: e034c13 ("rescan-scsi-bus: improve handling of USB SD card readers")